requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-ratelimit
python-decouple
Pillow
argon2-cffi
razorpay
requests
python-slugify
bleach
cryptography
